pyproject.toml
[tool.poetry]
name = "kicad-symbol-gen"
version = "0.1.0"
description = ""
authors = ["Jiajie Chen <c@jia.je>"]

[tool.poetry.dependencies]
python = "^3.8"
sexpdata = "^0.0.3"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
